// packages/note-lib/src/enums/IntervalQualityEnums.ts
export enum IntervalQualityName {
    Perfect = "Perfect",
    Major = "Major",
    Minor = "Minor",
    Augmented = "Augmented",
    Diminished = "Diminished",
    Suspended = "Suspended",
    Dominant = "Dominant",
    Flat = "Flat",
    Sharp = "Sharp",
    HalfDiminished = "Half Diminished",
}

export enum IntervalQualityShortHand {
    Perfect = "P",
    Major = "M",
    Minor = "m",
    Augmented = "A",
    Diminished = "d",
    Suspended = "sus",
    Dominant = "dom",
    maj = "maj",
    min = "min",
    aug = "aug",
    dim = "dim",
}

export enum IntervalQualitySymbol {
    Major = "△",
    Augmented = "+",
    Raised = "+",
    Lowered = "-",
    Diminished = "°",
    HalfDiminished = "ø",
    Flat = "b",
    Sharp = "#",
    FlatSymbol = "♭",
    SharpSymbol = "♯",
}